src/prediction/requirements.txt
fastapi
fastapi-health
uvicorn
sentry_sdk
newrelic
python-dotenv
requests
openai